# api-doc-dist/swagger.json
{
  "components": {
    "schemas": {
...
      "MemberModel": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "isFavorite": {
            "type": "boolean",
            "description": "If `true`, the user has saved this trip for future reference",
            "enum": [true, false]
          },
          "isMember": {
            "type": "boolean",
            "description": "Is member of the trip, if `true` then counted as active member of the trip",
            "enum": [true, false]
          },
          "isActive": {
            "type": "boolean",
            "description": "If `isMember=true` then this value applies",
            "enum": [true, false]
          },
          "isOwner": {
            "type": "boolean",
            "description": "Member is owner of the trip or not",
            "enum": [true, false]
          },
          "updatedAt": {
            "type": "number",
            "format": "Unix timestamp",
            "description": "The last time user interacted with the trip"
          }
        }
      },
...
      "CommonErrorFormat": {
        "type": "object",
        "required": ["status", "result"],
        "properties": {
          "status": {
            "type": "string",
            "default": "error"
          },
          "result": {
            "type": "object",
            "required": ["errorCode", "message"],
            "properties": {
              "errorCode": {
                "type": "string",
                "description": "Error code to handle internationalization(i18n)"
              },
              "message": {
                "type": "string",
                "description": "Error message in english"
              },
              "field": {
                "type": "string",
                "description": "Error field name"
              }
            }
          }
        }
      },
      "CommonSuccessFormat": {
        "type": "object",
        "required": ["status", "result"],
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          },
          "result": {
            "type": "string",
            "default": "success"
          }
        }
      },
...
      "TripModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "number",
            "format": "YYYYMMDD"
          },
          "endDate": {
            "type": "number",
            "format": "YYYYMMDD"
          },
          "focus": {
            "type": "string"
          },
          "destinations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minGroupSize": {
            "type": "number"
          },
          "maxGroupSize": {
            "type": "number"
          },
          "spotsAvailable": {
            "type": "number"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["english"]
            }
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pictureUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "itineraries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "desciption": {
                  "type": "string"
                },
                "imageUrl": {
                  "type": "string"
                }
              }
            }
          },
          "rooms": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "cost": {
                  "type": "number"
                },
                "available": {
                  "type": "number"
                },
                "filled": {
                  "type": "number"
                }
              }
            }
          },
          "addOns": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "cost": {
                  "type": "number"
                },
                "available": {
                  "type": "number"
                },
                "filled": {
                  "type": "number"
                }
              }
            }
          },
          "isDepositApplicable": {
            "type": "boolean"
          },
          "deposit": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number"
              },
              "expirationDate": {
                "type": "number"
              },
              "includeAddOns": {
                "type": "boolean"
              }
            }
          },
          "isDiscountApplicable": {
            "type": "boolean"
          },
          "discounts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "discType": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "expirationDate": {
                  "type": "number"
                },
                "includeAddOns": {
                  "type": "boolean"
                }
              }
            }
          },
          "priceIncludes": {
            "type": "string"
          },
          "priceExcludes": {
            "type": "string"
          },
          "lastBookingDate": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "published", "completed", "cancelled"]
          },
          "tripPaymentType": {
            "type": "string",
            "enum": ["pay", "payasyougo", "free"]
          },
          "tripLength": {
            "type": "number"
          },
          "groupSize": {
            "type": "number"
          },
          "spotsFilled": {
            "type": "number"
          },
          "favoriteCount": {
            "type": "number"
          },
          "spotsReserved": {
            "type": "number"
          },
          "isLocked": {
            "type": "boolean"
          },
          "showGroupHistory": {
            "type": "boolean"
          },
          "isPublic": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "isFull": {
            "type": "boolean"
          },
          "isArchived": {
            "type": "boolean",
            "description": "Trip is expired, startDate less than todays date."
          },
          "isFavorite": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "number",
            "format": "unix timestamp"
          },
          "updatedAt": {
            "type": "number",
            "format": "unix timestamp"
          },
          "ownerId": {
            "type": "string"
          },
          "updatedBy": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "pathTripId": {
        "name": "tripId",
        "in": "path",
        "description": "Id of the trip",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "queryMemberId": {
        "name": "memberId",
        "in": "query",
        "description": "Id of the user",
        "schema": {
          "type": "string"
        }
      },
...
      "queryPage": {
        "name": "page",
        "in": "query",
        "description": "Page number of pagination",
        "schema": {
          "type": "number",
          "default": 0
        }
      },
      "queryMatchExact": {
        "name": "matchExactDate",
        "in": "query",
        "description": "Match `startDate` and `endDate` exact",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "queryLimit": {
        "name": "limit",
        "in": "query",
        "description": "Limit of the records on a each page",
        "schema": {
          "type": "number",
          "default": 25,
          "minimum": 25,
          "maximum": 100
        }
      },
      "querySortBy": {
        "name": "sortBy",
        "in": "query",
        "description": "Sort by",
        "schema": {
          "type": "string",
          "default": "updatedAt",
          "enum": ["updatedAt"]
        }
      },
...
      "querySortOrder": {
        "name": "sortOrder",
        "in": "query",
        "description": "Sort order applied on `sortBy` fields",
        "schema": {
          "type": "number",
          "enum": [1, -1],
          "default": -1
        }
      },
      "querySearch": {
        "name": "search",
        "in": "query",
        "description": "Search filter",
        "schema": {
          "type": "string"
        }
      },
      "queryMinGroupSize": {
        "name": "minGroupSize",
        "in": "query",
        "description": "Search for minGroupSize",
        "schema": {
          "type": "number"
        }
      },
      "queryMaxGroupSize": {
        "name": "maxGroupSize",
        "in": "query",
        "description": "Search for maxGroupSize",
        "schema": {
          "type": "number"
        }
      },
      "queryMinStartDate": {
        "name": "startDate",
        "in": "query",
        "description": "Search for minStartDate. Format: `YYYYMMDD`",
        "schema": {
          "type": "number",
          "format": "YYYYMMDD"
        }
      },
      "queryMaxEndDate": {
        "name": "endDate",
        "in": "query",
        "description": "Search for maxEndDate. Format: `YYYYMMDD`",
        "schema": {
          "type": "number",
          "format": "YYYYMMDD"
        }
      },
      "queryMinTripLength": {
        "name": "minTripLength",
        "in": "query",
        "description": "Search for minTripLength",
        "schema": {
          "type": "number"
        }
      },
      "queryMaxTripLength": {
        "name": "maxTripLength",
        "in": "query",
        "description": "Search for maxTripLength",
        "schema": {
          "type": "number"
        }
      },
      "queryInterests": {
        "name": "interests",
        "in": "query",
        "description": "Search for interests, comma separated values. Ex. value1,value2",
        "schema": {
          "type": "string"
        }
      },
      "queryDestinations": {
        "name": "destinations",
        "in": "query",
        "description": "Search for destinations, comma separated values. Ex. value1,value2",
        "schema": {
          "type": "string"
        }
      },
      "queryIsMember": {
        "name": "isMember",
        "in": "query",
        "description": "List all the trips belong to user",
        "schema": {
          "type": "boolean"
        }
      },
      "queryIsFavorite": {
        "name": "isFavorite",
        "in": "query",
        "description": "List all the favorite trips of the user",
        "schema": {
          "type": "boolean"
        }
      },
      "queryIsPublic": {
        "name": "isPublic",
        "in": "query",
        "description": "List all the public trips of the user",
        "schema": {
          "type": "boolean",
          "default": true
        }
      },
      "queryIsArchived": {
        "name": "isArchived",
        "in": "query",
        "description": "List all the archived trips of the user",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "queryStatus": {
        "name": "status",
        "in": "query",
        "description": "List all the trips with status",
        "schema": {
          "type": "string",
          "enum": ["draft", "published", "completed", "cancelled"]
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    },
    "requestBodies": {
...
      "createTrip": {
        "description": "Create trip payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateUpdateTrip"
            }
          }
        }
      },
      "updateTrip": {
        "description": "Update trip payload",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateUpdateTrip"
            }
          }
        }
      },
...
    }
  },
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0",
    "title": "Tripsha API documentation",
    "description": "This is complete API documentation of all the API",
    "termsOfService": "https://tripsha.com/terms/",
    "contact": {
      "name": "Tripsha API Team",
      "email": "madanesanjay@yahoo.com",
      "url": "https://tripsha.com"
    }
  },
  "servers": [
    {
      "url": "https://staging-api.tripsha.com"
    }
  ],
  "paths": {
...
    "/trips/{tripId}/members": {
      "get": {
        "description": "Returns list of trip members",
        "operationId": "listMembers",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTripId"
          },
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Members response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "required": ["totalCount", "count", "data"],
                      "properties": {
                        "totalCount": {
                          "type": "number",
                          "description": "Total number of records"
                        },
                        "count": {
                          "type": "number",
                          "description": "Current number of records"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MemberModel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
...
    "/trips": {
      "get": {
        "description": "Returns list of trips",
        "operationId": "listTrips",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          },
          {
            "$ref": "#/components/parameters/querySortBy"
          },
          {
            "$ref": "#/components/parameters/querySortOrder"
          },
          {
            "$ref": "#/components/parameters/queryMinGroupSize"
          },
          {
            "$ref": "#/components/parameters/queryMaxGroupSize"
          },
          {
            "$ref": "#/components/parameters/queryMinStartDate"
          },
          {
            "$ref": "#/components/parameters/queryMatchExact"
          },
          {
            "$ref": "#/components/parameters/queryMaxEndDate"
          },
          {
            "$ref": "#/components/parameters/queryMinTripLength"
          },
          {
            "$ref": "#/components/parameters/queryMaxTripLength"
          },
          {
            "$ref": "#/components/parameters/queryInterests"
          },
          {
            "$ref": "#/components/parameters/queryDestinations"
          }
        ],
        "responses": {
          "200": {
            "description": "Trip response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "number",
                          "description": "Current records count"
                        },
                        "total": {
                          "type": "number",
                          "description": "Total number of records."
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TripModel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a trip",
        "operationId": "createTrip",
        "tags": ["Trips"],
        "requestBody": {
          "$ref": "#/components/requestBodies/createTrip"
        },
        "responses": {
          "200": {
            "description": "Trip response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "result": {
                      "type": "string",
                      "description": "Trip id"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
    "/mytrips": {
      "get": {
        "description": "Returns list of my trips",
        "operationId": "myTrips",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          },
          {
            "$ref": "#/components/parameters/querySortBy"
          },
          {
            "$ref": "#/components/parameters/querySortOrder"
          },
          {
            "$ref": "#/components/parameters/querySearch"
          },
          {
            "$ref": "#/components/parameters/queryIsMember"
          },
          {
            "$ref": "#/components/parameters/queryIsFavorite"
          },
          {
            "$ref": "#/components/parameters/queryIsPublic"
          },
          {
            "$ref": "#/components/parameters/queryIsArchived"
          },
          {
            "$ref": "#/components/parameters/queryStatus"
          }
        ],
        "responses": {
          "200": {
            "description": "Trip response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "description": "Total number of records"
                        },
                        "count": {
                          "type": "number",
                          "description": "Current records count."
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TripModel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/saved-trips": {
      "get": {
        "description": "Returns list of saved trips",
        "operationId": "savedTrips",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/queryPage"
          },
          {
            "$ref": "#/components/parameters/queryLimit"
          },
          {
            "$ref": "#/components/parameters/querySortBy"
          },
          {
            "$ref": "#/components/parameters/querySortOrder"
          },
          {
            "$ref": "#/components/parameters/querySearch"
          },
          {
            "$ref": "#/components/parameters/queryIsMember"
          },
          {
            "$ref": "#/components/parameters/queryIsFavorite"
          },
          {
            "$ref": "#/components/parameters/queryMemberId"
          },
          {
            "$ref": "#/components/parameters/queryIsArchived"
          }
        ],
        "responses": {
          "200": {
            "description": "Trip response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "totalCount": {
                          "type": "number",
                          "description": "Total number of records"
                        },
                        "count": {
                          "type": "number",
                          "description": "Current records count."
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TripModel"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
    "/trips/{tripId}": {
      "get": {
        "description": "Returns a trip based on a trip ID",
        "operationId": "findTripById",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTripId"
          }
        ],
        "responses": {
          "200": {
            "description": "Trip response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "result"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "default": "success"
                    },
                    "result": {
                      "$ref": "#/components/schemas/TripModel"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Updates a single trip based on the ID supplied and data",
        "operationId": "updateTrip",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTripId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/updateTrip"
        },
        "responses": {
          "204": {
            "description": "Trip updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "deletes a single trip based on the ID supplied",
        "operationId": "deleteTrip",
        "tags": ["Trips"],
        "parameters": [
          {
            "$ref": "#/components/parameters/pathTripId"
          }
        ],
        "responses": {
          "204": {
            "description": "Trip deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonSuccessFormat"
                }
              }
            }
          },
          "default": {
            "description": "Error Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorFormat"
                }
              }
            }
          }
        }
      }
    },
...
  }
}